+++ common/src/main/java/com/evandev/reliable_requiem/modules/ModEffects.java
@@ -2,21 +2,20 @@ package com.evandev.reliable_requiem.modules;
 
 import com.evandev.reliable_requiem.Constants;
 import com.evandev.reliable_requiem.effect.MementoMoriEffect;
-import net.minecraft.core.Holder;
-import net.minecraft.core.Registry;
-import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.resources.ResourceLocation;
+import com.evandev.reliable_requiem.registration.util.RegistrationProvider;
+import com.evandev.reliable_requiem.registration.util.RegistryObject;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.world.effect.MobEffect;
 
 public class ModEffects {
-    public static Holder<MobEffect> MEMENTO_MORI;
 
-    public static void init() {
-        MobEffect effect = Registry.register(
-                BuiltInRegistries.MOB_EFFECT,
-                ResourceLocation.fromNamespaceAndPath(Constants.MOD_ID, "memento_mori"),
-                new MementoMoriEffect()
-        );
-        MEMENTO_MORI = BuiltInRegistries.MOB_EFFECT.wrapAsHolder(effect);
+    public static final RegistrationProvider<MobEffect> MOB_EFFECTS = RegistrationProvider.get(Registries.MOB_EFFECT, Constants.MOD_ID);
+
+    public static final RegistryObject<MobEffect> MEMENTO_MORI = MOB_EFFECTS.register(
+            "memento_mori",
+            MementoMoriEffect::new
+    );
+
+    public static void load() {
     }
 }