+++ fabric/src/main/java/com/evandev/reliable_requiem/client/ModMenuIntegration.java
@@ -1,12 +1,17 @@
 package com.evandev.reliable_requiem.client;
 
 import com.evandev.reliable_requiem.client.integration.ClothConfigIntegration;
+import com.evandev.reliable_requiem.platform.Services;
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
+import net.fabricmc.loader.api.FabricLoader;
 
 public class ModMenuIntegration implements ModMenuApi {
     @Override
     public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return ClothConfigIntegration::createScreen;
+        if (Services.PLATFORM.isModLoaded("cloth-config")) {
+            return ClothConfigIntegration::createScreen;
+        }
+        return null;
     }
 }