+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2026-08-06
+
+### Added
+
+- Added configurable health loss upon death.
+- Added Crystal Shards and Crystal Hearts for restoring health, inspired by Matcha Flavoured.
+- Added more config options related to health loss/restoration.
+
+### Changed
+
+- Changed `spread` config options from a toggle to a multiplier.
+
+### Fixed
+
+- Fixed conditional cloth config registration on Fabric.
+
 ## [1.1.1] - 2026-07-10
 
 ### Fixed